bpf: common: untangle bpf_sock_term.c from lib/common.h
Shift a few definitions around so that we don't need to include the whole
common.h header.
