+++ bpf/bpf_alignchecker.c
@@ -49,10 +49,13 @@ add_type(struct drop_notify);
 add_type(struct policy_verdict_notify);
 add_type(struct debug_msg);
 add_type(struct debug_capture_msg);
+
+#include "lib/sock.h"
 add_type(struct ipv4_revnat_tuple);
 add_type(struct ipv4_revnat_entry);
 add_type(struct ipv6_revnat_tuple);
 add_type(struct ipv6_revnat_entry);
+
 add_type(struct ipv4_frag_id);
 add_type(struct ipv4_frag_l4ports);
 add_type(struct ipv6_frag_id);
